other companies: don't sort by contribution count
The 'other companies' search card is used on the browse companies page.
The sorting in the search takes 7 seconds and we do it twice - for the list and for the count.
Until we have a better solution I think we better give up the sorting and save 14 seconds loading time.